requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
environs==9.5.0
idna==3.3
iniconfig==1.1.1
marshmallow==3.17.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
